Make Adjacent internal
